core/src/main/java/com/avolution/actor/ActorSystem.java
package com.avolution.actor;

import java.util.concurrent.ConcurrentHashMap;
import java.util.concurrent.ConcurrentMap;
import java.util.concurrent.ExecutorService;
import java.util.concurrent.Executors;

public class ActorSystem {
    private final ConcurrentMap<String, AbstractActor> actorRegistry;
    private final ExecutorService executorService;

    public ActorSystem() {
        this.actorRegistry = new ConcurrentHashMap<>();
        this.executorService = Executors.newCachedThreadPool();
    }

    public AbstractActor createActor(Class<? extends AbstractActor> actorClass) {
        try {
            AbstractActor actor = actorClass.getDeclaredConstructor().newInstance();
            String actorId = generateUniqueId();
            actor.setId(actorId);
            actorRegistry.put(actorId, actor);
            return actor;
        } catch (Exception e) {
            throw new RuntimeException("Failed to create actor", e);
        }
    }

    public AbstractActor getActor(String actorId) {
        return actorRegistry.get(actorId);
    }

    public void terminateActor(String actorId) {
        AbstractActor actor = actorRegistry.remove(actorId);
        if (actor != null) {
            actor.stop();
        }
    }

    public void sendMessage(String actorId, Message message) {
        AbstractActor actor = getActor(actorId);
        if (actor != null) {
            actor.receiveMessage(message);
        }
    }

    public void addChildActor(String parentId, String childId) {
        AbstractActor parent = getActor(parentId);
        AbstractActor child = getActor(childId);
        if (parent != null && child != null) {
            parent.addChild(child);
        }
    }

    public void removeChildActor(String parentId, String childId) {
        AbstractActor parent = getActor(parentId);
        AbstractActor child = getActor(childId);
        if (parent != null && child != null) {
            parent.removeChild(child);
        }
    }

    public void restartActor(String actorId) {
        AbstractActor actor = getActor(actorId);
        if (actor != null) {
            actor.restart();
        }
    }

    public void stopActor(String actorId) {
        AbstractActor actor = getActor(actorId);
        if (actor != null) {
            actor.stop();
        }
    }

    public void resumeActor(String actorId) {
        AbstractActor actor = getActor(actorId);
        if (actor != null) {
            actor.resume();
        }
    }

    public void monitorActorHealth(String actorId) {
        AbstractActor actor = getActor(actorId);
        if (actor != null) {
            // Implement health monitoring logic here
        }
    }

    public void handleActorFailure(String actorId, Throwable cause) {
        AbstractActor actor = getActor(actorId);
        if (actor != null) {
            actor.propagateError(cause);
        }
    }

    public void traverseHierarchy() {
        for (AbstractActor actor : actorRegistry.values()) {
            traverseActorHierarchy(actor);
        }
    }

    private void traverseActorHierarchy(AbstractActor actor) {
        // Implement hierarchy traversal logic here
    }

    public void distributeActor(String actorId, String nodeId) {
        AbstractActor actor = getActor(actorId);
        if (actor != null) {
            // Implement actor distribution logic here
        }
    }

    public void replicateActorState(String actorId) {
        AbstractActor actor = getActor(actorId);
        if (actor != null) {
            // Implement actor state replication logic here
        }
    }

    public void monitorDistributedSystem() {
        // Implement distributed system monitoring logic here
    }

    private String generateUniqueId() {
        return java.util.UUID.randomUUID().toString();
    }
}
