core/src/test/java/com/avolution/actor/ActorSystemTest.java
package com.avolution.actor;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class ActorSystemTest {

    private ActorSystem actorSystem;

    @BeforeEach
    void setUp() {
        actorSystem = new ActorSystem();
    }

    @Test
    void testCreateActor() {
        AbstractActor actor = actorSystem.createActor(BasicActor.class);
        assertNotNull(actor);
        assertNotNull(actorSystem.getActor(actor.getId()));
    }

    @Test
    void testRetrieveActor() {
        AbstractActor actor = actorSystem.createActor(BasicActor.class);
        AbstractActor retrievedActor = actorSystem.getActor(actor.getId());
        assertEquals(actor, retrievedActor);
    }

    @Test
    void testTerminateActor() {
        AbstractActor actor = actorSystem.createActor(BasicActor.class);
        actorSystem.terminateActor(actor.getId());
        assertNull(actorSystem.getActor(actor.getId()));
    }

    @Test
    void testSendMessage() {
        AbstractActor actor = actorSystem.createActor(BasicActor.class);
        Message message = new Message("Hello");
        actorSystem.sendMessage(actor.getId(), message);
        // Add assertions to verify message sending
    }

    @Test
    void testAddChildActor() {
        AbstractActor parent = actorSystem.createActor(BasicActor.class);
        AbstractActor child = actorSystem.createActor(BasicActor.class);
        actorSystem.addChildActor(parent.getId(), child.getId());
        assertTrue(parent.getChildren().contains(child));
        assertEquals(parent, child.getParent());
    }

    @Test
    void testRemoveChildActor() {
        AbstractActor parent = actorSystem.createActor(BasicActor.class);
        AbstractActor child = actorSystem.createActor(BasicActor.class);
        actorSystem.addChildActor(parent.getId(), child.getId());
        actorSystem.removeChildActor(parent.getId(), child.getId());
        assertFalse(parent.getChildren().contains(child));
        assertNull(child.getParent());
    }

    @Test
    void testRestartActor() {
        AbstractActor actor = actorSystem.createActor(BasicActor.class);
        actorSystem.restartActor(actor.getId());
        // Add assertions to verify actor restart
    }

    @Test
    void testStopActor() {
        AbstractActor actor = actorSystem.createActor(BasicActor.class);
        actorSystem.stopActor(actor.getId());
        // Add assertions to verify actor stop
    }

    @Test
    void testResumeActor() {
        AbstractActor actor = actorSystem.createActor(BasicActor.class);
        actorSystem.resumeActor(actor.getId());
        // Add assertions to verify actor resume
    }

    @Test
    void testMonitorActorHealth() {
        AbstractActor actor = actorSystem.createActor(BasicActor.class);
        actorSystem.monitorActorHealth(actor.getId());
        // Add assertions to verify actor health monitoring
    }

    @Test
    void testHandleActorFailure() {
        AbstractActor actor = actorSystem.createActor(BasicActor.class);
        Throwable cause = new RuntimeException("Test error");
        actorSystem.handleActorFailure(actor.getId(), cause);
        // Add assertions to verify actor failure handling
    }

    @Test
    void testTraverseHierarchy() {
        actorSystem.traverseHierarchy();
        // Add assertions to verify hierarchy traversal
    }

    @Test
    void testDistributeActor() {
        AbstractActor actor = actorSystem.createActor(BasicActor.class);
        actorSystem.distributeActor(actor.getId(), "node1");
        // Add assertions to verify actor distribution
    }

    @Test
    void testReplicateActorState() {
        AbstractActor actor = actorSystem.createActor(BasicActor.class);
        actorSystem.replicateActorState(actor.getId());
        // Add assertions to verify actor state replication
    }

    @Test
    void testMonitorDistributedSystem() {
        actorSystem.monitorDistributedSystem();
        // Add assertions to verify distributed system monitoring
    }
}
